split per-Schema conversion into its own file
